NEXUS-7789: Moved components into nexus-scheduling-api and renamed the bundle
