+++ components/nexus-scheduling/src/main/java/org/sonatype/nexus/scheduling/TaskDescriptorSupport.java
