+++ components/nexus-scheduling/src/main/java/org/sonatype/nexus/scheduling/TaskInfo.java
