+++ components/nexus-scheduling/src/main/java/org/sonatype/nexus/scheduling/internal/DefaultTaskFactory.java
@@ -11,7 +11,7 @@
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
 
-package org.sonatype.nexus.internal.scheduling;
+package org.sonatype.nexus.scheduling.internal;
 
 import java.util.List;
 import java.util.Map;
