+++ components/nexus-scheduling/src/main/java/org/sonatype/nexus/scheduling/internal/ThreadPoolTaskExecutorSPI.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.internal.scheduling;
+package org.sonatype.nexus.scheduling.internal;
 
 import java.util.Date;
 import java.util.List;
