+++ components/nexus-scheduling/src/test/java/org/sonatype/nexus/scheduling/internal/TaskSchedulerTest.java
@@ -10,14 +10,14 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.internal.scheduling;
+package org.sonatype.nexus.scheduling.internal;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.Future;
 
 import javax.inject.Named;
 
-import org.sonatype.nexus.internal.scheduling.Tasks.SleeperTask;
+import org.sonatype.nexus.scheduling.internal.Tasks.SleeperTask;
 import org.sonatype.nexus.scheduling.Task;
 import org.sonatype.nexus.scheduling.TaskConfiguration;
 import org.sonatype.nexus.scheduling.TaskDescriptor;
