+++ components/pom.xml
@@ -43,7 +43,7 @@
     <module>nexus-ehcache</module>
     <module>nexus-extender</module>
     <module>nexus-httpclient-api</module>
-    <module>nexus-scheduling-api</module>
+    <module>nexus-scheduling</module>
     <module>nexus-formfields</module>
     <module>nexus-launcher</module>
     <module>nexus-locks</module>
