+++ pom.xml
@@ -258,7 +258,7 @@
 
       <dependency>
         <groupId>org.sonatype.nexus</groupId>
-        <artifactId>nexus-scheduling-api</artifactId>
+        <artifactId>nexus-scheduling</artifactId>
         <version>3.0.0-SNAPSHOT</version>
       </dependency>
